requirements.txt
Flask
Pillow
numpy
opencv-python
tensorflow
flask-cors